// src/app/model/cliente.ts
export class cliente{
    id?: number;
    nombre: string;
    apellido: string;
    fechaAlta: Date;
    cuil?: string;
    cuit: string;
    telefono1:string;
    telefono2?: string;
    email: string;
    sitioWeb?: string;
    razonSocial?: string;
    direccion: string;
    localidad: string;
    provincia: string;
    codigoPostal: string;

}